Separate platforms code to different directories
This loses the files' history, use `git log --follow` for more details.
