+++ windows/windows.py
@@ -28,7 +28,7 @@ def GrabImage():
     timestamp = time.strftime("%y%m%d-%H%M%S")
     img_name = "PastedImage" + timestamp + ".png"
     file_path = (
-        bpy.context.preferences.addons[__package__].preferences.default_img_dir
+        bpy.context.preferences.addons["ImagePaste"].preferences.default_img_dir
         + img_name
     )
